Sort new m2m fields and adjust indentation
New Many-to-many fields in models are now sorted, this is done in order to maintain their order consistent with the old m2m fields. This will improve the readability and understanding of the model changes. In addition, Adjusted the indentation throughout the file for better readability and coding standard conformity.